refactor components module
