+++ src/App.js
@@ -1,6 +1,6 @@
 import "./css/App.css";
-import Timer from "./Timer";
-import Settings from "./Settings";
+import Timer from "./components/Timer";
+import Settings from "./components/Settings";
 import { useState } from "react";
 import SettingsContext from "./SettingsContext";
